change int to byte in getInput/OutputCount

+++ src/main/java/vitzli/miscmfrcircuits/circuits/analog/MulDiv.java
@@ -10,12 +10,12 @@ public class MulDiv extends StatelessCircuit {
 
     
     @Override
-    public int getInputCount() {
+    public byte getInputCount() {
         return 3;
     }
 
     @Override
-    public int getOutputCount() {
+    public byte getOutputCount() {
         return 3;
     }
