+++ src/main/java/vitzli/miscmfrcircuits/circuits/logicboolean/DAC.java
@@ -5,12 +5,12 @@ import vitzli.miscmfrcircuits.circuits.base.StatelessCircuit;
 public class DAC extends StatelessCircuit {
 
 	@Override
-	public int getInputCount() {
+	public byte getInputCount() {
 		return 16;
 	}
 
 	@Override
-	public int getOutputCount() {
+	public byte getOutputCount() {
 		return 1;
 	}
